requirements.txt
# --- Manipulación de PDFs ---
pymupdf==1.23.6
pdfminer.six==20221105
pikepdf==8.7.1
pdfplumber==0.10.3

# --- OCR y Procesamiento de Imágenes ---
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0

# --- Validación y análisis de estructura ---
lxml==5.1.0
jsonschema==4.21.1
colorcontrast==1.0.5

# --- Interfaz Gráfica (Qt) ---
PySide6==6.6.2
qdarkstyle==3.0.3
qtawesome==1.3.0
qdarktheme==1.1.1

# --- Generación de informes ---
jinja2==3.1.3
weasyprint==61.2
matplotlib==3.8.4
markdown2==2.4.10

# --- Utilidades generales ---
pyyaml==6.0.1
loguru==0.7.2
watchdog==4.0.0

pip install pymupdf==1.23.6 pdfminer.six==20221105 pikepdf==8.7.1 pdfplumber==0.10.3 pytesseract==0.3.10 opencv-python==4.9.0.80 Pillow==10.2.0 lxml==5.1.0 jsonschema==4.21.1 colorcontrast==1.0.5 PySide6==6.6.2 qdarkstyle==3.0.3 qtawesome==1.3.0 qdarktheme==1.1.1 jinja2==3.1.3 weasyprint==61.2 matplotlib==3.8.4 markdown2==2.4.10 pyyaml==6.0.1 loguru==0.7.2 watchdog==4.0.0
pip install langcodes
